Rename actor file
